refactor: rename fsutil to filesystem and split diagnostics
- rename the `fsutil` module to `filesystem` (the abbreviation violated the
  no-abbreviation naming rule; the module now holds only the capped reader).
- split `compose/diagnostics.rs` (474 lines) into a `diagnostics/` directory:
  the parsed-but-unsupported field collectors move to `ignored_fields.rs`,
  leaving the unknown-key passes and `collect` in `mod.rs` (305 lines).

Refs #279 (file-length hygiene). Predicate-bool field renames and the
remaining near-cap splits are deferred — the bool renames touch public
library types (RunOptions/DevelopConfig) and need owner-coordinated breaking
changes.
